# backend/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: NyumbaLink API
  version: "2.0.0"
  description: |
    ## NyumbaLink — AI-Powered Accommodation Brokerage System
    **Dar es Salaam, Tanzania**

    NyumbaLink connects **Seekers** (tenants) with verified **Landlords** across
    Dar es Salaam's five municipalities — Kinondoni, Ilala, Ubungo, Temeke, and Kigamboni.

    ### Authentication
    Protected endpoints require a **Bearer JWT** in the `Authorization` header.
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.
    Access tokens expire after **24 hours**.
    Use `POST /api/v1/auth/refresh` to renew.

    ### Roles
    | Role | Description |
    |---|---|
    | `SEEKER` | Tenant searching for accommodation |
    | `LANDLORD` | Property owner managing listings |
    | `ADMIN` | Platform administrator |

    ### Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.nyumbalink.co.tz`
  contact:
    name: NyumbaLink Support
    email: mnemosadick@gmail.com
    url: https://nyumbalink.co.tz
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.nyumbalink.co.tz
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via `POST /api/v1/auth/login`.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    AmenityOut:
      type: object
      properties:
        id:
          type: integer
          example: 1
        amenity_name:
          type: string
          nullable: true
          example: "Swimming Pool"
    AmenityCreate:
      type: object
      required: [amenity_name]
      properties:
        amenity_name:
          type: string
          example: "Gym"
    PropertyImageOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_url:
          type: string
          example: "https://res.cloudinary.com/nyumbalink/image/upload/v1234/prop.jpg"
        is_primary:
          type: boolean
          example: true
    PropertyVideoOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_url:
          type: string
          nullable: true
          example: "https://res.cloudinary.com/nyumbalink/video/upload/v1234/tour.mp4"
    LandlordBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
          nullable: true
    PropertyOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        landlord_id:
          type: string
          format: uuid
        landlord:
          $ref: "#/components/schemas/LandlordBrief"
        property_type:
          type: string
          enum: [HOSTEL, ROOM, APARTMENT, HOUSE]
          example: "HOSTEL"
        title:
          type: string
          nullable: true
          example: "Kiota Hostel (Girls Only)"
        description:
          type: string
          nullable: true
          example: "Premium girls-only hostel in Mikocheni with 175 beds."
        location:
          type: string
          nullable: true
          example: "Mikocheni, Kinondoni"
        ward:
          type: string
          nullable: true
          example: "Mikocheni"
        district:
          type: string
          nullable: true
          example: "Kinondoni"
        city:
          type: string
          example: "Dar es Salaam"
        latitude:
          type: number
          format: float
          nullable: true
          example: -6.7658
        longitude:
          type: number
          format: float
          nullable: true
          example: 39.2795
        monthly_rent:
          type: number
          format: float
          nullable: true
          example: 640000
        bedrooms:
          type: integer
          nullable: true
          example: 1
        bathrooms:
          type: integer
          nullable: true
          example: 24
        size_sqm:
          type: number
          format: float
          nullable: true
          example: 18.0
        gender_restriction:
          type: string
          enum: [MALE, FEMALE, MIXED]
          nullable: true
          example: "FEMALE"
        capacity:
          type: integer
          nullable: true
          example: 175
        availability_status:
          type: string
          enum: [AVAILABLE, RENTED, SUSPENDED, MAINTENANCE]
          example: "AVAILABLE"
        is_verified:
          type: boolean
          example: true
        view_count:
          type: integer
          example: 142
        images:
          type: array
          items:
            $ref: "#/components/schemas/PropertyImageOut"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/PropertyVideoOut"
        amenities:
          type: array
          items:
            $ref: "#/components/schemas/AmenityOut"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropertyCreate:
      type: object
      required: [property_type, title, monthly_rent]
      properties:
        property_type:
          type: string
          enum: [HOSTEL, ROOM, APARTMENT, HOUSE]
          example: "ROOM"
        title:
          type: string
          minLength: 3
          maxLength: 255
          example: "Spacious Room in Mikocheni B"
        description:
          type: string
          nullable: true
          example: "Clean, quiet room with 24/7 water and LUKU electricity."
        location:
          type: string
          nullable: true
          example: "Mikocheni B, Off New Bagamoyo Road"
        ward:
          type: string
          nullable: true
          example: "Mikocheni"
        district:
          type: string
          nullable: true
          example: "Kinondoni"
        city:
          type: string
          default: "Dar es Salaam"
        latitude:
          type: number
          format: float
          nullable: true
          example: -6.764
        longitude:
          type: number
          format: float
          nullable: true
          example: 39.279
        monthly_rent:
          type: number
          format: float
          minimum: 1
          example: 180000
        bedrooms:
          type: integer
          nullable: true
          minimum: 0
          example: 1
        bathrooms:
          type: integer
          nullable: true
          minimum: 0
          example: 1
        size_sqm:
          type: number
          format: float
          nullable: true
          example: 18.0
        gender_restriction:
          type: string
          enum: [MALE, FEMALE, MIXED]
          nullable: true
        capacity:
          type: integer
          nullable: true
          example: 1
        amenity_ids:
          type: array
          items:
            type: integer
          default: []
          example: [1, 3, 5]
    PropertyUpdate:
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        ward:
          type: string
          nullable: true
        monthly_rent:
          type: number
          format: float
          nullable: true
          minimum: 1
        bedrooms:
          type: integer
          nullable: true
        bathrooms:
          type: integer
          nullable: true
        size_sqm:
          type: number
          nullable: true
        gender_restriction:
          type: string
          nullable: true
        capacity:
          type: integer
          nullable: true
        availability_status:
          type: string
          enum: [AVAILABLE, RENTED, SUSPENDED, MAINTENANCE]
          nullable: true
        amenity_ids:
          type: array
          items:
            type: integer
          nullable: true
    PaginatedProperties:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/PropertyOut"
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 3

    # ── Rentals ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/properties:
    get:
      tags: [Properties]
      summary: Search / list properties
      description: |
        Full-text and filtered property search.
        All query parameters are optional — omitting them returns all available properties.
      security: []
      operationId: properties_list
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Free-text search (title, description, location)
          example: "mikocheni hostel"
        - name: ward
          in: query
          schema:
            type: string
          example: "Mikocheni"
        - name: district
          in: query
          schema:
            type: string
          example: "Kinondoni"
        - name: property_type
          in: query
          schema:
            type: string
            enum: [HOSTEL, ROOM, APARTMENT, HOUSE]
          example: "HOSTEL"
        - name: min_rent
          in: query
          schema:
            type: number
            minimum: 0
          example: 100000
        - name: max_rent
          in: query
          schema:
            type: number
            minimum: 0
          example: 800000
        - name: min_bedrooms
          in: query
          schema:
            type: integer
            minimum: 0
          example: 1
        - name: gender_restriction
          in: query
          schema:
            type: string
            enum: [MALE, FEMALE, MIXED]
          description: Filter hostels by gender policy
          example: "FEMALE"
        - name: availability
          in: query
          schema:
            type: string
            enum: [AVAILABLE, RENTED, SUSPENDED, MAINTENANCE]
            default: AVAILABLE
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated property list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProperties"
    post:
      tags: [Properties]
      summary: Create a new property listing
      description: |
        Create a new listing. Requires **LANDLORD** or **ADMIN** role.
        `amenity_ids` links the property to existing amenities.
      operationId: properties_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyCreate"
            examples:
              hostel:
                summary: Create a hostel listing
                value:
                  property_type: "HOSTEL"
                  title: "Dage Hostel (Girls Only)"
                  description: "Girls-only hostel in Africa Sana with 56-bed capacity."
                  location: "Africa Sana, Kinondoni"
                  ward: "Sinza"
                  district: "Kinondoni"
                  latitude: -6.7955
                  longitude: 39.2468
                  monthly_rent: 640000
                  bedrooms: 1
                  bathrooms: 8
                  gender_restriction: "FEMALE"
                  capacity: 56
                  amenity_ids: [1, 2]
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyOut"
        "403":
          description: Not a landlord or admin
  /api/v1/properties/my/listings:
    get:
      tags: [Properties]
      summary: Get my listings (Landlord)
      description: Returns all properties owned by the authenticated landlord.
      operationId: properties_my_listings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of landlord's properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProperties"
...
  /api/v1/properties/amenities:
    post:
      tags: [Properties]
      summary: Create a new amenity (Admin only)
      operationId: amenities_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmenityCreate"
      responses:
        "201":
          description: Amenity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmenityOut"
  /api/v1/properties/{property_id}:
    get:
      tags: [Properties]
      summary: Get a single property
      description: Returns full property detail including images, videos, and amenities. Also increments the view counter.
      security: []
      operationId: properties_get_one
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyOut"
        "404":
          description: Property not found
    patch:
      tags: [Properties]
      summary: Update a property
      description: Partial update. Landlord can only update own properties; Admin can update any.
      operationId: properties_update
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyOut"
    delete:
      tags: [Properties]
      summary: Suspend (soft-delete) a property
      description: Sets `availability_status` to `SUSPENDED`. Does not physically delete the record.
      operationId: properties_delete
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "204":
          description: Property suspended
  /api/v1/properties/{property_id}/images:
    post:
      tags: [Properties]
      summary: Add an image URL to a property
      operationId: properties_add_image
      parameters:
        - $ref: "#/components/parameters/PropertyId"
        - name: image_url
          in: query
          required: true
          schema:
            type: string
          example: "https://res.cloudinary.com/nyumbalink/image/upload/v1234/room.jpg"
        - name: is_primary
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "201":
          description: Image added
  /api/v1/properties/{property_id}/images/{image_id}:
    delete:
      tags: [Properties]
      summary: Delete a property image
      operationId: properties_delete_image
      parameters:
        - $ref: "#/components/parameters/PropertyId"
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Image deleted
  /api/v1/properties/{property_id}/videos:
    post:
      tags: [Properties]
      summary: Add a video URL to a property
      operationId: properties_add_video
      parameters:
        - $ref: "#/components/parameters/PropertyId"
        - name: video_url
          in: query
          required: true
          schema:
            type: string
          example: "https://res.cloudinary.com/nyumbalink/video/upload/v1234/tour.mp4"
      responses:
        "201":
          description: Video added

  # ────────────────────────────────── RENTALS ────────────────────────────────
...